web/requirements.txt
Django
gunicorn
psycopg2
djangorestframework
numpy
